backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
jwt==1.3.1
Markdown==3.5.2
pycparser==2.21
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
